Fixing the camera side
Basically scrapped using CameraUIMode wasn't sure about its
compatibility with Windows IOT, but it was working on my PC but not on
the PI.

This uses a capture element to actually stream before I take a photo and
show the preview on the side of the preview. Still needs alot of ui
work, but I'll learn that xaml some other time.