UI,Calls and Storage problem solved
So the calls are more efficient, instead of printing them out to a seperate file, I just use the built in calls from the ProjectOxford namespace. Should've done that from the start.grrrr.
It no longer breaks when I feed it the random access stream. The trick was cloning the stream and not working off the existing one thats running because there'd be conflict. UI is nicer, about time, not the best but it'll do for now. Right now, I want to get the actual recomendation algorithm to work with facial hair (yes this is working now WOOO), age and gender, maybe emotion. If I get that done, I'll start porting it over to be pi compatible. then its gonna be designing the simple circuit. I'm pretty excited to finally get started on that!!!